Split production.rs into subdirectory modules
Refactor the 702-line production.rs into a production/ directory with
separate modules for better maintainability:

- types.rs (255 lines): IndustryType, GoodsType enums, ProductionChain,
  IndustryBuilding component, CityGoods resource
- systems.rs (370 lines): assign_industry_type, update_production_chains,
  and helper functions (extract_from_deposits, consume_goods, etc.)
- tests.rs (61 lines): unit tests
- mod.rs (23 lines): plugin definition and pub re-exports

All public API items are re-exported from mod.rs so external callers
are unaffected.

Closes #1389
